Ultratb is almost entirely refactored out of interactiveshell.
Generally the color scheme arguments are being invoked less to
classes SyntaxTB and AutoFormattedTB.

The InteractiveShell is using traceback module a lot more so that
bugs don't propagate as violently.

Unfortunately it still doesn't start. The old colorize module or
coloransi or whatever is difficult to reason about from the extreme
number of name changes in the variables. There's either going to be some
extreme incompatabilities or someone a lot smarter than I is going to need to
figure out how to rework this.

OR.

We could just start using `self` for our instances? :)