Continued reorganizing the docs. Really excited.
The docs are now built by Sphinx and the makefile.

No need for autogen scripts. And as a product of simple reorganization,
they look so overwhelmingly long, I'd be surprised to hear anybody
asking for them back.

Sphinx indexes things in non-intuitive ways, so this layout should
make it way easier for new users to navigate the site.

In addition:

- Ran black over interactiveshell

- Moved IPython/testing/plugin back to it's original location.
  Apparently the tests depend on them being there.