Review what was failing in tests
Added a handful of `get_ipython` calls. May seem unnecessary
as our tests are already set up to handle this; however, in an attempt
to reduce the amount of global state that every module depends on for
execution, it may be beneficial to either explicitly state the package
fixtures or drill them in 1 by 1.