Move external/qt_for_kernel globals into ifmain.
Pytest keeps crashing because I don't have all the imports but
QT is ~90MB and I don't want to install it in a venv to get
pytest running.

Delete functions that we rewrote because updates to inspect made
their way into upstream years ago.

Still trying to whittle down unnecessary imports. The package
is so interconnected that bugs keep propagating everywhere and slowing
things down.

magics.execution still had an try/except for the python2 version of
cProfile. We don't maintain py2 compatability so deleted that too.